refactor: use @commander-js/extra-typings
This commit removes our own type definitions and utilizes '@commander-js/extra-typings' to organize the option-related types better. Because of this all custom validation code was rendered obsolete and was consequently removed.
The primary objective behind these changes is to keep the codebase more maintainable and efficient in terms of usage of external dependencies that are better designed for the task at hand.